+++ src/main/java/com/smunity/api/global/config/PasswordEncoderConfiguration.java
@@ -1,6 +1,6 @@
-package com.smunity.api.config;
+package com.smunity.api.global.config;
 
-import com.smunity.api.config.security.encoder.Pbkdf2PasswordEncoder;
+import com.smunity.api.global.config.security.encoder.Pbkdf2PasswordEncoder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.crypto.password.PasswordEncoder;
