+++ src/main/java/com/smunity/api/global/config/security/SecurityConfiguration.java
@@ -1,4 +1,4 @@
-package com.smunity.api.config.security;
+package com.smunity.api.global.config.security;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
