+++ src/main/java/com/smunity/api/global/config/security/encoder/Pbkdf2PasswordEncoder.java
@@ -1,4 +1,4 @@
-package com.smunity.api.config.security.encoder;
+package com.smunity.api.global.config.security.encoder;
 
 import com.google.common.base.Charsets;
 import lombok.val;
