+++ src/test/java/com/smunity/api/config/security/encoder/Pbkdf2PasswordEncoderTest.java
@@ -1,5 +1,6 @@
 package com.smunity.api.config.security.encoder;
 
+import com.smunity.api.global.config.security.encoder.Pbkdf2PasswordEncoder;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
